+++ tests/integration/test_ai.py
@@ -32,8 +32,10 @@ def test_ai_spans_example_transaction(
     outcomes_consumer = outcomes_consumer()
 
     project_id = 42
-    mini_sentry.add_full_project_config(project_id)
-
+    project = mini_sentry.add_full_project_config(project_id)
+    project["config"].setdefault("features", []).extend(
+        ["organizations:relay-generate-billing-outcome"]
+    )
     mini_sentry.global_config["aiModelMetadata"] = {
         "version": 1,
         "models": {
@@ -1296,8 +1298,29 @@ def test_ai_spans_example_transaction(
         },
     ]
 
+    num_messages = 2
     if relay_emits_accepted_outcome:
-        assert outcomes_consumer.get_aggregated_outcomes() == [
+        num_messages = 3
+    outcomes = outcomes_consumer.get_aggregated_outcomes(n=num_messages)
+
+    if relay_emits_accepted_outcome:
+        assert outcomes == [
+            {
+                "category": DataCategory.TRANSACTION.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
+            {
+                "category": DataCategory.SPAN.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 10,
+            },
             {
                 "category": DataCategory.SPAN_INDEXED.value,
                 "key_id": 123,
@@ -1305,5 +1328,24 @@ def test_ai_spans_example_transaction(
                 "outcome": 0,
                 "project_id": 42,
                 "quantity": 10,
-            }
+            },
+        ]
+    else:
+        assert outcomes == [
+            {
+                "category": DataCategory.TRANSACTION.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 1,
+            },
+            {
+                "category": DataCategory.SPAN.value,
+                "key_id": 123,
+                "org_id": 1,
+                "outcome": 0,
+                "project_id": 42,
+                "quantity": 10,
+            },
         ]